# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Legend Academy API Documentation",
        "description": "Legend Pay API description",
        "contact": {
            "email": "o.oyedele@suburbanfiberco.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Returns user data",
                "operationId": "Registration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Registration Successful"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "description": "Returns user data",
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "example@gmail.com"
                                    },
                                    "password": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Login Successful"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LoginResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Full Name of the new user",
                        "type": "string",
                        "example": "Jame Prince"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email address of the new user",
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the new user",
                        "type": "string",
                        "example": "12345678"
                    },
                    "phone_number": {
                        "title": "Phone Number",
                        "description": "Phone of the new User",
                        "type": "string",
                        "example": "0816155633"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Address of the new user",
                        "type": "string",
                        "example": "25 uloho street"
                    },
                    "user_type": {
                        "title": "User Type",
                        "description": "User Type of the new user",
                        "type": "string",
                        "example": "User"
                    },
                    "profile_image": {
                        "title": "Profile Image",
                        "description": "Profile Image of the new user",
                        "type": "string"
                    },
                    "department_id": {
                        "title": "Department ID",
                        "description": "Department's id of the new user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "StoreUserRequest": {
                "title": "Store User request",
                "description": "Store User request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Full Name of the new user",
                        "type": "string",
                        "example": "Jame Prince"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email address of the new user",
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the new user",
                        "type": "string",
                        "example": "12345678"
                    },
                    "phone_number": {
                        "title": "Phone Number",
                        "description": "Phone of the new User",
                        "type": "string",
                        "example": "0816155633"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Address of the new user",
                        "type": "string",
                        "example": "25 uloho street"
                    },
                    "user_type": {
                        "title": "User Type",
                        "description": "User Type of the new user",
                        "type": "string",
                        "example": "User"
                    },
                    "profile_image": {
                        "title": "Profile Image",
                        "description": "Profile Image of the new user",
                        "type": "string"
                    },
                    "department_id": {
                        "title": "Department ID",
                        "description": "Department's id of the new user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "LoginResource": {
                "title": "LoginResource",
                "description": "Login resource",
                "properties": {
                    "token": {
                        "title": "token",
                        "description": "access token",
                        "type": "string",
                        "example": "fjfjfjfjfuf74fjffhfhfhshgfaga"
                    },
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Full Name of the new user",
                        "type": "string",
                        "example": "Jame Prince"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email address of the new user",
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "phone_number": {
                        "title": "Phone Number",
                        "description": "Phone of the new User",
                        "type": "string",
                        "example": "0816155633"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Address of the new user",
                        "type": "string",
                        "example": "25 uloho street"
                    },
                    "user_type": {
                        "title": "User Type",
                        "description": "User Type of the new user",
                        "type": "string",
                        "example": "User"
                    },
                    "profile_image": {
                        "title": "Profile Image",
                        "description": "Profile Image of the new user",
                        "type": "string"
                    },
                    "department_id": {
                        "title": "Department ID",
                        "description": "Department's id of the new user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginResource"
                }
            },
...
        }
    }
}